refactoring io.chino.examples.utils to io.chino.examples
