+++ src/main/java/io/chino/examples/auth/AuthSamples.java
@@ -10,7 +10,7 @@ import io.chino.api.schema.SchemaRequest;
 import io.chino.api.schema.SchemaStructure;
 import io.chino.api.user.User;
 import io.chino.api.userschema.UserSchema;
-import io.chino.examples.util.DeleteAll;
+import io.chino.examples.DeleteAll;
 import io.chino.java.ChinoAPI;
 import io.chino.examples.userschemas.UserSchemaStructureSample;
 import io.chino.test.util.Constants;
